requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pandas==2.2.1
google-cloud-storage==2.16.0
python-dotenv==1.0.1
requests==2.31.0
setuptools>=67.0.0  # Speeds up dependency resolution
